fixed bugs:
-- miss member value pointer
-- resize by reconctruct
